+++ lib/enigma.rb
@@ -1,10 +1,4 @@
 class Enigma
-  attr_reader :today, :character_set
-  
-  def initialize
-    @today = "date"
-    @character_set = ("a".."z").to_a << " "
-  end
   
   def encrypt(message, key, date)
     {encryption: "keder ohulw", key: "02715", date: "040895"}
